copy received YAML/JSON data into an internal format
This makes it possible to change the data format sent by Gatemons without
affecting the Javascript code for the website.

The internal format is mostly similar to the current Gatemon JSON format,
but adds some sub-structures for easier processing in app.js, and removes
some unnecessary sub-structures.

The entire code for checking and copying the received data has been
moved to the new sanitizeYamlJsonInput() function.
To simplify error handling, this function uses PHP Exceptions if an
error occurs, and leaves it to the caller to actually send an HTTP
error to the Gatemon.